projects:ad9434-fmc-500ebz: Use iio_app
Refactor project to use common iio_app
