+++ projects/ad9434-fmc-500ebz/src/parameters.h
@@ -48,6 +48,7 @@
 #define ADC_DDR_BASEADDR			XPAR_DDR_MEM_BASEADDR + 0x800000
 #define UART_DEVICE_ID				XPAR_XUARTPS_0_DEVICE_ID
 #define UART_IRQ_ID				XPAR_XUARTPS_1_INTR
+#define UART_BAUDRATE                           115200
 #define INTC_DEVICE_ID				XPAR_SCUGIC_SINGLE_DEVICE_ID
 
 #endif /* _PARAMETERS_H_ */
